\name{BarleyPheno}
\alias{BarleyPheno}
\title{Phenotype data from 194 phenotype trials}
\description{
The data was downloaded from <https://triticeaetoolbox.org/barley/> and curated. The data is in list format. Each element in the list is data from one phenotypic trial concerning a sample of traits. 
}
\usage{
data(BarleyPheno)
}
