% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomb.R
\name{coef.covcomb}
\alias{coef.covcomb}
\title{Extract covariance matrix from covcomb object}
\usage{
\method{coef}{covcomb}(object, ...)
}
\arguments{
\item{object}{An object of class \code{covcomb}}

\item{...}{Additional arguments (unused)}
}
\value{
The estimated covariance matrix
}
\description{
Extract covariance matrix from covcomb object
}
