% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomb.R
\name{compute_aic}
\alias{compute_aic}
\title{Compute Akaike Information Criterion (AIC) for a fitted covcomb model}
\usage{
compute_aic(fit)
}
\arguments{
\item{fit}{A fitted object of class "covcomb" from fit_covcomb()}
}
\value{
Numeric value of AIC
}
\description{
Compute Akaike Information Criterion (AIC) for a fitted covcomb model
}
\details{
AIC is computed as -2 * log_likelihood + 2 * k, where k is the number of
free parameters (unique elements in the covariance matrix).
}
\examples{
\dontrun{
fit <- fit_covcomb(S_list, nu_vec)
aic <- compute_aic(fit)
}
}
