% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomb.R
\name{compute_bic}
\alias{compute_bic}
\title{Compute Bayesian Information Criterion (BIC) for a fitted covcomb model}
\usage{
compute_bic(fit, nu)
}
\arguments{
\item{fit}{A fitted object of class "covcomb" from fit_covcomb()}

\item{nu}{Named numeric vector of degrees of freedom used in fitting}
}
\value{
Numeric value of BIC
}
\description{
Compute Bayesian Information Criterion (BIC) for a fitted covcomb model
}
\details{
BIC is computed as -2 * log_likelihood + log(n) * k, where k is the number
of free parameters and n is the total sample size (sum of degrees of freedom).
}
\examples{
\dontrun{
fit <- fit_covcomb(S_list, nu_vec)
bic <- compute_bic(fit, nu_vec)
}
}
