% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_methods.R
\name{compute_Q_function}
\alias{compute_Q_function}
\title{Compute Q Function}
\usage{
compute_Q_function(theta, theta_ref, internal_data, ridge = 1e-08)
}
\arguments{
\item{theta}{Parameters at which to evaluate Q}

\item{theta_ref}{Reference parameters for taking expectation}

\item{internal_data}{Preprocessed data structure}

\item{ridge}{Ridge for numerical stability}
}
\value{
Scalar value of Q function
}
\description{
Computes the Q function Q(theta | theta_ref) = E[complete log-lik | obs, theta_ref]
This is the expected complete-data log-likelihood.
}
\keyword{internal}
