% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_methods.R
\name{compute_se_plugin}
\alias{compute_se_plugin}
\title{Plugin Standard Errors}
\usage{
compute_se_plugin(sigma_hat, coverage_mat)
}
\arguments{
\item{sigma_hat}{Estimated covariance matrix}

\item{coverage_mat}{Matrix of coverage weights (sum of degrees of freedom
for pairs of variables observed together)}
}
\value{
Matrix of standard errors with the same dimensions as sigma_hat
}
\description{
Computes "plugin" standard errors assuming complete data and using the
closed-form Wishart variance with coverage weights.
}
\keyword{internal}
