% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{.is_free_model}
\alias{.is_free_model}
\title{Is This n_factors Encoding the Free-Sigma Model?}
\usage{
.is_free_model(n_factors)
}
\arguments{
\item{n_factors}{Value using \code{fit_covcomb}'s encoding.}
}
\value{
\code{TRUE} for \code{NULL} and \code{"free"}.
}
\description{
Is This n_factors Encoding the Free-Sigma Model?
}
\keyword{internal}
