% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{.k_max}
\alias{.k_max}
\title{Saturation Point for the Number of Factors}
\usage{
.k_max(p)
}
\arguments{
\item{p}{Number of variables.}
}
\value{
Largest k beyond which a k-factor model adds no parameters relative
  to the free-\eqn{\Sigma} model.
}
\description{
Saturation Point for the Number of Factors
}
\keyword{internal}
