% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomb.R
\name{fitted.covcomb}
\alias{fitted.covcomb}
\title{Fitted covariance matrix for covcomb object}
\usage{
\method{fitted}{covcomb}(object, ...)
}
\arguments{
\item{object}{An object of class \code{covcomb}}

\item{...}{Additional arguments (unused)}
}
\value{
The combined covariance estimate on the aggregated scale (S_hat)
}
\description{
Fitted covariance matrix for covcomb object
}
