% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_methods.R
\name{jacobian_logchol_to_sigma}
\alias{jacobian_logchol_to_sigma}
\title{Jacobian of Sigma w.r.t. log-Cholesky parameters}
\usage{
jacobian_logchol_to_sigma(theta)
}
\arguments{
\item{theta}{Vector of log-Cholesky parameters}
}
\value{
Jacobian matrix with dimensions (p^2) x length(theta)
}
\description{
Computes the Jacobian matrix d(vec(Sigma))/d(theta) for the
log-Cholesky parameterization using numerical differentiation.
}
\keyword{internal}
