% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_methods.R
\name{logchol_to_sigma}
\alias{logchol_to_sigma}
\title{Inverse Log-Cholesky Parameterization}
\usage{
logchol_to_sigma(theta)
}
\arguments{
\item{theta}{Vector of parameters: log(diagonals) followed by
free off-diagonals (below diagonal)}
}
\value{
Positive definite covariance matrix
}
\description{
Converts log-Cholesky parameters back to a covariance matrix.
Reconstructs lower-triangular L such that Sigma = L L^T.
}
\keyword{internal}
