% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomb.R
\name{print.covcomb}
\alias{print.covcomb}
\title{Print method for covcomb objects}
\usage{
\method{print}{covcomb}(x, ...)
}
\arguments{
\item{x}{An object of class \code{covcomb}}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for covcomb objects
}
