% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability.R
\name{print.covcomb_identifiability}
\alias{print.covcomb_identifiability}
\title{Print an Identifiability Report}
\usage{
\method{print}{covcomb_identifiability}(x, max_pairs = 10L, ...)
}
\arguments{
\item{x}{An object of class \code{covcomb_identifiability}.}

\item{max_pairs}{Maximum number of unidentified pairs to list individually.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an Identifiability Report
}
