% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_methods.R
\name{sigma_to_logchol}
\alias{sigma_to_logchol}
\title{Log-Cholesky Parameterization}
\usage{
sigma_to_logchol(Sigma)
}
\arguments{
\item{Sigma}{Positive definite covariance matrix}
}
\value{
Vector of parameters: log(diagonals) followed by free
  off-diagonals (below diagonal)
}
\description{
Converts a covariance matrix to log-Cholesky parameters.
Uses lower-triangular Cholesky decomposition: Sigma = L L^T where
L is lower triangular.
}
\keyword{internal}
