% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covcomb.R
\name{summary.covcomb}
\alias{summary.covcomb}
\title{Summary method for covcomb objects}
\usage{
\method{summary}{covcomb}(object, ...)
}
\arguments{
\item{object}{An object of class \code{covcomb}}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Summary method for covcomb objects
}
