\name{CovCombR-package}
\alias{CovCombR-package}
\alias{CovCombR}
\docType{package}
\title{
\packageTitle{CovCombR}
}
\description{
\packageDescription{CovCombR}
}
\details{
The input to the main program CovComb is a list of partial covariance matrices. The output is an estimated combined (high dimensional) covariance matrix. The output of the algorithm, the completed covariance matrix, can be used to make inferences about unobserved covariances, as an input to sparse covariance estimation algorithms, in covariance graph estimation, in discriminant analysis.
}
\author{
\packageAuthor{CovCombR}

Maintainer: \packageMaintainer{CovCombR}
}
\references{
Adventures in Multi-Omics I: Combining heterogeneous data sets via relationships matrices Deniz Akdemir, Julio Isidro Sanchez. <https://www.biorxiv.org/content/10.1101/857425v1>.
}

\concept{covariance}
\concept{wishart distribution}
\concept{multiview data}
\concept{multiomics}
\concept{phenomics}
\concept{graphical models}
\concept{genomics}
\concept{covariance graph}
